Formalized two-way data transfer
Now the arduino data gets saved to a Map called data, from which we can
get any of it by saying data.get(key).

We need to decide on exactly what keys we will use, and their data
types. (Right now it only takes integers, but we can change that if we
deem it necessary).

Serial format Arduino -> Pi will be: "KEY VAL; KEY VAL"
and Pi -> Arduino will be: "KEYVAL;KEYVAL;" (no spaces)

FirefighterMain is the main program which will put together all the
other parts. It has the public static void main function which controls
the whole program.