Testing 03/03/2015
If the sensor reading is bigger than 60 AND we expect the sensor reading
to be bigger than sixty, boost that guess and move on with your life.

Robot estimates pose most of the time. I'm starting to thing that a
combination of this estimate with a small-scale localization (i.e.
taking the sensor data literally rather than mixing it
probabilistically) will be helpful.

Promising, but still scarily far from finished.